requirements.txt
# TeraBox Downloader Bot Requirements
# Python 3.9+

# Core Dependencies
python-telegram-bot==20.3
telegram==0.0.1

# Async HTTP Client
aiohttp==3.9.1
aiofiles==23.2.1

# Database
motor==3.3.2
pymongo==4.6.1

# Image Processing
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
requests==2.31.0

# Logging & Monitoring
colorlog==6.8.0

# Development (optional)
black==23.12.0
flake8==6.1.0
pylint==3.0.3
pytest==7.4.3
pytest-asyncio==0.21.1
